#4362 - Move suggestion extraction from RecommendationService into separate class
- Moved the code out
